refactor: split MavenCentralUploaderPlugin into configurators
